✨ Add proper shop items distribution (#38)
* Fix generating cards inside shop

* Add proper shop and booster pack cards editions, enhancements and seals distribution

* Fix Aura spectral card edition distribution

* Add proper booster pack type and size distribution

* Add guaranteed normal buffoon pack in the first shop

* Add proper shop items distribution including vouchers and ghost deck related weight changes

* Fix The Soul and Black Hole spectral cards appearing in shop

* Add The Soul and Black Hole to spectral packs with proper odds

* Remove outdated notes about not implemented vouchers

* Add proper joker weights

* Fix base edition card weight

+++ content/tarot.c
@@ -215,7 +215,7 @@ uint8_t use_tarot_card(Tarot tarot) {
       if (!random_chance(1, 4)) break;
 
       // Foil, Holographic, Polychrome
-      uint8_t edition_weights[] = {50, 35, 15};
+      uint16_t edition_weights[] = {50, 35, 15};
       state.game.jokers.cards[joker_index].edition = random_weighted(edition_weights, 3) + 1;
       break;
     }
